htpdate: use -x option rather than -s
With htpdate, if you start it up with "-s" and "-D", it will
daemonize and switch back to "-a".  That is, it will Daemonize,
and then try to slowly adjust the clock rather than setting it
immediately.

This is not what we want, because the Raspberry Pi starts up like
it's 2016.  The kernel rejects the call to `adjtime()`.

After investigating the source, `-x` appears to do what we want
rather than slowly changing it.

Signed-off-by: Sean Cross <sean@xobs.io>